Create Yyd
Artist painting on the outside wall.  Pyramids characters equipment, graffiti effect.  The artist wore a black jacket with the word “BADR” written on it, jeans, and white shoes.  Near the artist are cans of spray paint and other drawing equipment.  The background of the photo shows a wet city street.
    Regarding the subject of the mural, the idea was modified again.  Here I present an example of the character of Pharaoh